refactor: move all remaining vendored deps into stacks_common, since signers will need them too
