+++ stacks-common/src/deps_common/mod.rs
@@ -16,6 +16,7 @@
 
 pub mod bech32;
 pub mod bitcoin;
+pub mod ctrlc;
 pub mod httparse;
 
 #[cfg(test)]
